Changed the layout and functionality of the Midi Note Choosing Page
Changed the functionality and the entire layout of the MidiNotesChoosingPage. It used to be plenty of buttons and combobox menus, each for a different sound, now it's one combobox responsible for choosing the sound, another one for choosing a MIDI note for the corresponding sound, and one big button for triggering the sound that's being chosen at the moment. Such a change helps with performance (the GUI loads faster) and is more future-proof (easier to add new sounds without being limited to the space provided in GUI).