Changed the way of loading samples
The samples can now be loaded from binary data embedded in the plugin. This eliminates the need to read files from disk space, which prevents situations in which the user might accidentally remove some of the wav files. This change should also make the loading of samples a little bit quicker. Currently only electronic samples are loaded (only 30 samples), hopefully stuffing the plugin with all the samples won't make it load for ages.