GUI changes
Added the last backkgroud image for the acoustic drum set. Changed the size of the entire GUI to 1050 x 500 (because there has been a new cymbal added on the right, and it overlapped with the TabbedComponent). It used to be 1000 x 500. Added button for Immersive sound in the acoustic drum kit.