Loading samples finished
Finished all the code responsible for loading samples. Added a new functionality that loads "essential samples" from binary data if one of the wav files in a specified destination is missing. Couldn't load all samples from binary data because there was not enough memory on my laptop for the compiler to compile the file properly and also the plugin would be pretty big in size (it's already 142 MB).